Add support for specifying a specific Git ref and commit changes automatically
This commit adds the necessary code to support specifying a specific Git ref using the new `--ref` option. If the `--ref` option is specified, the `get_diff` function will now use the `ref` argument to specify the range of commits to diff. Additionally, the `--commit` option has been modified to automatically commit the changes to the repository using the `git apply --cached` command. If the `--commit` option is specified, the diff will be applied using `git apply --cached` and the return code will be checked. If the `git apply --cached` command fails, the output will be printed.